Rename TodoActions.js to index.js
