+++ src/components/AddTodo.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { addTodo } from '../actions/TodoActions'
+import { addTodo } from '../actions'
 
 let AddTodo = ({ dispatch }) => {
   let input;
